fix wrong bulma class for notifications; rename component file to a more appropriately named one; don't subscribe if no rpc endpoint is selected, show an error message if so
