+++ webui/src/components/app-notifications.vue
@@ -1,29 +1,23 @@
-
 <script setup lang="ts">
-import {useNotificationsStore} from '@/stores/notifications';
-import {Status} from '@/lib/status';
+import { useNotificationsStore } from '@/stores/notifications';
+import { Status } from '@/lib/status';
 
 const store = useNotificationsStore();
 </script>
 
 <template>
   <div class="notifications-wrapper">
-    <div
-      v-for="notif in store.notifications"
-      :key="notif.message"
-      class="notification"
-      :class="{
-        'is-success': notif.status === Status.Success,
-        'is-warn': notif.status === Status.Warn,
-        'is-error': notif.status === Status.Error,
-      }"
-    >
-      {{notif.message}}
+    <div v-for="notif in store.notifications" :key="notif.message" class="notification" :class="{
+      'is-success': notif.status === Status.Success,
+      'is-warning': notif.status === Status.Warn,
+      'is-danger': notif.status === Status.Error,
+    }">
+      {{ notif.message }}
     </div>
   </div>
 </template>
 
-<style  scoped >
+<style scoped>
 .notifications-wrapper {
   display: flex;
   flex-direction: column;
@@ -36,9 +30,8 @@ const store = useNotificationsStore();
   color: #fdf9f9;
   font-weight: 500;
 
-  &.is-success{
+  &.is-success {
     background-color: #539953;
   }
 }
-
 </style>
